refactor(borrower): split monolith service into per-concern submodules
``app/services/borrower.py`` grew to ~1000 lines across the borrowers
epic (#221, #244 series). Time to split before the next feature lands.

Strict no-behavior-change refactor — every function moved verbatim,
the public import surface stays identical via re-exports.

## Layout

    backend/app/services/borrower/
    ├── __init__.py    # re-exports everything; ``__all__`` lists the
    │                   contract pinned by existing callers + tests
    ├── _common.py     # ANONYMIZED_BORROWER_NAME, BorrowerWithStats,
    │                   BorrowerStatsPage, BorrowerLoanRow,
    │                   _normalize_name, _normalize_contact
    ├── query.py       # list_borrowers, list_borrowers_with_stats,
    │                   list_loans_for_borrower, get_borrower_or_404,
    │                   get_borrower_detail_or_404, BorrowerStatusFilter
    ├── lifecycle.py   # create_borrower, update_borrower,
    │                   link_loans_to_borrowers
    ├── anonymize.py   # _hard_anonymize_in_memory, anonymize_borrower,
    │                   restore_borrower, bulk_anonymize_borrowers,
    │                   select_borrowers_for_retention_anonymize,
    │                   bulk_anonymize_borrowers_by_inactivity,
    │                   finalize_due_pending_anonymizations,
    │                   ANONYMIZE_PENDING_TTL
    └── merge.py       # MergeResult, MERGE_UNDO_TTL,
                        _snapshot_borrower, _restore_borrower_from_snapshot,
                        _hash_undo_token, merge_borrowers,
                        apply_merge_undo, gc_expired_merge_undo_logs

## Public API stability

All historical imports work unchanged:

    from app.services.borrower import (
        anonymize_borrower,
        ANONYMIZED_BORROWER_NAME,
        _normalize_name,  # also re-exported — test_borrower_audit.py uses it
        ...
    )

``__all__`` in ``__init__.py`` lists everything, including the three
underscore-prefixed names that tests pin contracts on. ``api/borrowers.py``,
``api/loan_service.py``, and the seven borrower test files needed
zero changes.

## Why this split

- ``query`` has no mutating dependencies — easy to read top-down.
- ``anonymize`` and ``merge`` carry the most state (pending TTL,
  undo log) and benefit from being read in isolation.
- ``lifecycle`` is small (create/update/link); kept separate from
  ``query`` so the read-only file stays read-only.

## Verification (local)

- ruff: clean
- mypy: clean (129 files checked)
- ``pytest tests/test_borrower_*.py tests/test_borrowers.py``: 107
  passed on sqlite
